Api for deleting post

+++ controllers/api/v1/post-api.js
@@ -1,33 +1,33 @@
-const Post=require('../../../models/post');
-const Comment=require('../../../models/comments');
+const Post = require('../../../models/post');
+const Comment = require('../../../models/comments');
 
-module.exports.index= async function(req,res){
-    try{
+module.exports.index = async function (req, res) {
+    try {
 
         // finding all posts along with comments and users name
-    let posts = await Post.find({})
-    .sort('-createdAt')
-    .populate('user')
-    .populate({
-        path: 'comment',
-        populate: {
-            path: 'user'        
-        }
-    })
+        let posts = await Post.find({})
+            .sort('-createdAt')
+            .populate('user')
+            .populate({
+                path: 'comment',
+                populate: {
+                    path: 'user'
+                }
+            })
 
-    return res.json(200,{
-        message:"List of Posts",
-        posts:posts
-    });
-}
+        return res.json(200, {
+            message: "List of Posts",
+            posts: posts
+        });
+    }
 
-    catch(err){
+    catch (err) {
 
-      console.log("****** Error",err);
+        console.log("****** Error", err);
 
-      return res.json(500,{
-          message:"Internal Server Error"
-      });
+        return res.json(500, {
+            message: "Internal Server Error"
+        });
 
     }
 }
@@ -35,24 +35,32 @@ module.exports.index= async function(req,res){
 
 // Deleting posts through Api's
 
-module.exports.destroy =async function (req, res) {
+module.exports.destroy = async function (req, res) {
 
-    try{
-        let post=await Post.findById(req.params.id); 
+    try {
+        let post = await Post.findById(req.params.id);
         //here authentication is done post.user check the user id from post model
-        //if (post.user == req.user.id) {    // .id means converting the object id into string
+        if (post.user == req.user.id) {    // .id means converting the object id into string
             post.remove();
 
             await Comment.deleteMany({ post: req.params.id });
-        
-                return res.json(200,{
-                    message:"Post and Comments Deleted"
-                });
-    }catch(err){
-        console.log("****** Error",err);
 
-        return res.json(500,{
-            message:"Internal Server Error"
+            return res.json(200, {
+                message: "Post and Comments Deleted"
+            });
+        } else {
+            return res.json(401, {
+                message: "You cannot delete this Post"
+            });
+
+
+        }
+    }
+    catch (err) {
+        console.log("****** Error", err);
+
+        return res.json(500, {
+            message: "Internal Server Error"
         });
 
     }
